server/app/core/config.py
"""
This configuration file is responsible for retrieving environment settings and secrets.
"""

import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL", "mysql+pymysql://usuario:contraseña@localhost:3306/mi_db")
SECRET_KEY = os.getenv("SECRET_KEY", "a_secret_key")
ALGORITHM = os.getenv("ALGORITHM", "HS256")